requirements.txt
annotated-types==0.6.0
anyio==4.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openai==1.3.9
pydantic==2.5.2
pydantic-core==2.14.5
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
tiktoken==0.5.2
tqdm==4.66.1
typing-extensions==4.9.0
urllib3==2.1.0
